requirements.txt
botocore
boto3
testfixtures
formic-py3
nose
